+++ Correlator/Module/capture.py
@@ -1,7 +1,10 @@
+import logging
 from mako.template import Template
 
 from Correlator.event import NoticeEvent, EventProcessor, AuditEvent
 from Correlator.util import Module, format_timestamp, calculate_summary
+
+log = logging.getLogger(__name__)
 
 
 class CaptureStatsEvent(AuditEvent):
@@ -20,9 +23,8 @@ class CaptureStatsEvent(AuditEvent):
 
 class CaptureOnly(Module):
 
-    def __init__(self, processor: EventProcessor, log):
+    def __init__(self, processor: EventProcessor):
 
-        self.log = log
         self.processor = processor
         self.description = 'Syslog Capture support'
         self.identifier = 'Capture'
@@ -36,15 +38,11 @@ class CaptureOnly(Module):
         self.end = None
 
     def statistics(self, reset=False):
-        if self.start and self.end:
-            duration = (str(self.end - self.start))
-        else:
-            duration = None
 
         data = {
             'start': format_timestamp(self.start),
             'end': format_timestamp(self.end),
-            'duration': duration,
+            'duration': self._calculate_duration(self.start, self.end),
             'messages': self.num_records,
             'size': self.size_records
         }
@@ -66,9 +64,8 @@ class CaptureOnly(Module):
         if self.end is None or record.timestamp > self.end:
             self.end = record.timestamp
 
-        summary = calculate_summary('{} {} {} {}'.format(
-            record.hostname, record.appname, record.prog,
-            record.detail))
+        summary = calculate_summary(
+            f'{record.hostname} {record.appname} {record.prog} {record.detail}')
 
         self.dispatch_event(NoticeEvent(summary, record=record))
         self.num_records += 1